pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hftenv"
version = "0.1.0"
description = "HFT Security Environment for OpenEnv — high-frequency trading simulation"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "hftenv.server.app:main"

[tool.setuptools]
packages = ["hftenv", "hftenv.server"]
package-dir = {"hftenv" = ".", "hftenv.server" = "server"}
